+++ agr_indexer/src/main/java/org/alliancegenome/indexer/indexers/curation/service/AGMDiseaseAnnotationService.java
@@ -60,7 +60,7 @@ public class AGMDiseaseAnnotationService extends BaseDiseaseAnnotationService {
 					ret.add(da);
 				}
 			}
-			display.progressProcess(response.getReturnedRecords().longValue());
+			display.progressProcess(response.getReturnedRecords());
 		}
 		display.finishProcess();
 		writeToCache(cacheFileName, ret);
